refactor(app): close module card menu on click
This PR closes the module card menu when a menu item is clicked and also when the document is
clicked.

closes #9413
