+++ app/src/organisms/Devices/ModuleCard/ModuleOverflowMenu.tsx
@@ -1,59 +1,21 @@
 import * as React from 'react'
 import { useTranslation } from 'react-i18next'
 import { Flex, POSITION_RELATIVE } from '@opentrons/components'
-import {
-  THERMOCYCLER_MODULE_TYPE,
-  MAGNETIC_MODULE_TYPE,
-} from '@opentrons/shared-data'
 import { MenuList } from '../../../atoms/MenuList'
 import { MenuItem } from '../../../atoms/MenuList/MenuItem'
-import { MagneticModuleSlideout } from './MagneticModuleSlideout'
-import { TemperatureModuleSlideout } from './TemperatureModuleSlideout'
-import { ThermocyclerModuleSlideout } from './ThermocyclerModuleSlideout'
 
 import type { AttachedModule } from '../../../redux/modules/types'
 
 interface ModuleOverflowMenuProps {
   module: AttachedModule
+  handleClick: (isSecondary: boolean) => void
 }
 
 export const ModuleOverflowMenu = (
   props: ModuleOverflowMenuProps
 ): JSX.Element | null => {
   const { t } = useTranslation('device_details')
-  const { module } = props
-  const [showSlideout, setShowSlideout] = React.useState(false)
-  const [hasSecondary, setHasSecondary] = React.useState(false)
-
-  const renderSlideOut = (isSecondary: boolean = false): JSX.Element => {
-    if (module.type === THERMOCYCLER_MODULE_TYPE) {
-      return (
-        <ThermocyclerModuleSlideout
-          module={module}
-          onCloseClick={() => setShowSlideout(false)}
-          isExpanded={showSlideout}
-          isSecondaryTemp={isSecondary}
-        />
-      )
-    } else if (module.type === MAGNETIC_MODULE_TYPE) {
-      return (
-        <MagneticModuleSlideout
-          module={module}
-          onCloseClick={() => setShowSlideout(false)}
-          isExpanded={showSlideout}
-        />
-      )
-    } else {
-      return (
-        <TemperatureModuleSlideout
-          model={module.model}
-          serial={module.serial}
-          onCloseClick={() => setShowSlideout(false)}
-          isExpanded={showSlideout}
-        />
-      )
-    }
-  }
+  const { module, handleClick } = props
 
   const menuItems = {
     thermocyclerModuleType: [
@@ -84,13 +46,6 @@ export const ModuleOverflowMenu = (
     ],
   }
 
-  const handleClick = (isSecondary: boolean = false): void => {
-    if (isSecondary) {
-      setHasSecondary(true)
-    }
-    setShowSlideout(true)
-  }
-
   const AboutModuleBtn = (
     <MenuItem
       data-testid={`about_module_${module.model}`}
@@ -103,7 +58,6 @@ export const ModuleOverflowMenu = (
 
   return (
     <React.Fragment>
-      {showSlideout && renderSlideOut(hasSecondary)}
       <Flex position={POSITION_RELATIVE}>
         <MenuList
           buttons={[
