+++ app/src/organisms/Devices/ModuleCard/__tests__/ModuleOverflowMenu.test.tsx
@@ -8,24 +8,6 @@ import {
   mockThermocycler,
 } from '../../../../redux/modules/__fixtures__'
 import { ModuleOverflowMenu } from '../ModuleOverflowMenu'
-import { MagneticModuleSlideout } from '../MagneticModuleSlideout'
-import { TemperatureModuleSlideout } from '../TemperatureModuleSlideout'
-import { ThermocyclerModuleSlideout } from '../ThermocyclerModuleSlideout'
-
-jest.mock('../MagneticModuleSlideout')
-jest.mock('../TemperatureModuleSlideout')
-jest.mock('../ThermocyclerModuleSlideout')
-
-const mockMagneticModuleSlideout = MagneticModuleSlideout as jest.MockedFunction<
-  typeof MagneticModuleSlideout
->
-const mockTemperatureModuleSlideout = TemperatureModuleSlideout as jest.MockedFunction<
-  typeof TemperatureModuleSlideout
->
-
-const mockThermocyclerModuleSlideout = ThermocyclerModuleSlideout as jest.MockedFunction<
-  typeof ThermocyclerModuleSlideout
->
 
 const render = (props: React.ComponentProps<typeof ModuleOverflowMenu>) => {
   return renderWithProviders(<ModuleOverflowMenu {...props} />, {
@@ -38,16 +20,8 @@ describe('ModuleOverflowMenu', () => {
   beforeEach(() => {
     props = {
       module: mockMagneticModule,
+      handleClick: jest.fn(),
     }
-    mockMagneticModuleSlideout.mockReturnValue(
-      <div>Mock mag module slideout</div>
-    )
-    mockTemperatureModuleSlideout.mockReturnValue(
-      <div>Mock temperature module slideout</div>
-    )
-    mockThermocyclerModuleSlideout.mockReturnValue(
-      <div>Mock thermocycler module slideout</div>
-    )
   })
   afterEach(() => {
     jest.resetAllMocks()
@@ -61,7 +35,6 @@ describe('ModuleOverflowMenu', () => {
       background-color: transparent;
     `)
     fireEvent.click(buttonSetting)
-    getByText('Mock mag module slideout')
     const buttonAbout = getByRole('button', { name: 'About module' })
     fireEvent.click(buttonAbout)
     expect(buttonAbout).toBeEnabled()
@@ -82,13 +55,13 @@ describe('ModuleOverflowMenu', () => {
   it('renders the correct temperature module menu', () => {
     props = {
       module: mockTemperatureModuleGen2,
+      handleClick: jest.fn(),
     }
-    const { getByRole, getByText } = render(props)
+    const { getByRole } = render(props)
     const buttonSetting = getByRole('button', {
       name: 'Set module temperature',
     })
     fireEvent.click(buttonSetting)
-    getByText('Mock temperature module slideout')
     const buttonAbout = getByRole('button', { name: 'About module' })
     fireEvent.click(buttonAbout)
     expect(buttonAbout).toBeEnabled()
@@ -97,6 +70,7 @@ describe('ModuleOverflowMenu', () => {
   it('renders the correct TC module menu', () => {
     props = {
       module: mockThermocycler,
+      handleClick: jest.fn(),
     }
     const { getByRole } = render(props)
     const buttonSettingLid = getByRole('button', {
